fix(rest): Report not finding any objects when deleting by name
Previously we swallowed this case and made it seem like deletion (of nothing)
was successful.
As for the single delete case, it is not an 'error' when an object targeted for deletion
does not exist at all, as a delete.yaml is a growing list of configs that should not
exist - thus things that no longer exist might still be attempted for deletion.

+++ pkg/rest/config_upload.go
@@ -301,6 +301,12 @@ func deleteDynatraceObjects(client *http.Client, api api.Api, names []string, ur
 	}
 
 	var errors []error
+
+	if len(existingIds) == 0 {
+		log.Warn("No configs of given names found for deletion(%s): %s", api.GetId(), names)
+		return errors
+	}
+
 	for _, v := range existingIds {
 		if v.Id != "" {
 			err = deleteConfig(client, url, token, v.Id)
